Batter Stances Implemented
Player can use vertical keys to change their stance between lo, mid & hi
